@@ -1,42 +1,43 @@
 {
   "name": "ng-cache",
-  "version": "2.0.0",  
+  "version": "2.0.0",
+  "private": true,
   "author": "DagonMetric",
   "license": "MIT",
   "repository": {
     "type": "git",
-    "url": "https://github.com/dagonmetric/ng-cache.git"
+    "url": "https://github.com/Dagonmetric/ng-cache.git"
   },
   "bugs": {
-    "url": "https://github.com/dagonmetric/ng-cache/issues"
+    "url": "https://github.com/Dagonmetric/ng-cache/issues"
   },
-  "homepage": "https://github.com/dagonmetric/ng-cache",
+  "homepage": "https://github.com/Dagonmetric/ng-cache",
   "scripts": {
     "ng": "ng",
     "ngb": "ngb",
-    "rimraf": "rimraf",
     "lint": "ng lint",
-    "build": "ngb build --env.prod",    
+    "build": "ngb build --env.prod",
     "test": "ng test"
   },
   "dependencies": {
-    "@angular/common": "^8.0.0",
-    "@angular/core": "^8.0.0",
+    "@angular/common": "^8.1.0",
+    "@angular/core": "^8.1.0",
+    "@dagonmetric/ng-log": "^2.1.0",
     "rxjs": "^6.5.2",
     "tslib": "^1.10.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.800.2",
-    "@angular/cli": "^8.0.2",
-    "@angular/compiler": "^8.0.0",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/language-service": "^8.0.0",
-    "@angular/platform-browser": "^8.0.0",
-    "@angular/platform-browser-dynamic": "^8.0.0",
-    "@dagonmetric/angular-build": "^8.0.2",
+    "@angular-devkit/build-angular": "^0.801.0",
+    "@angular/cli": "^8.1.0",
+    "@angular/compiler": "^8.1.0",
+    "@angular/compiler-cli": "^8.1.0",
+    "@angular/language-service": "^8.1.0",
+    "@angular/platform-browser": "^8.1.0",
+    "@angular/platform-browser-dynamic": "^8.1.0",
+    "@dagonmetric/angular-build": "^8.0.4",
     "@types/jasmine": "^3.3.13",
     "@types/jasminewd2": "^2.0.6",
-    "@types/node": "^12.0.8",
+    "@types/node": "^12.0.10",
     "codelyzer": "^5.1.0",
     "jasmine-core": "^3.4.0",
     "jasmine-spec-reporter": "~4.2.1",
@@ -47,10 +48,10 @@
     "karma-jasmine-html-reporter": "^1.4.2",
     "karma-junit-reporter": "^1.2.0",
     "protractor": "~5.4.2",
-    "puppeteer": "^1.17.0",
-    "ts-node": "^8.2.0",
+    "puppeteer": "^1.18.1",
+    "ts-node": "^8.3.0",
     "tsickle": "^0.35.0",
-    "tslint": "^5.17.0",
+    "tslint": "^5.18.0",
     "tslint-microsoft-contrib": "^6.2.0",
     "typescript": "~3.4.5",
     "zone.js": "^0.9.1"
