@@ -22,13 +22,13 @@ module.exports = function (config) {
         coverageIstanbulReporter: {
             dir: require('path').join(__dirname, '../../dist/coverage/ng-cache'),
             reports: ['html', 'lcovonly', 'text-summary', 'cobertura'],
-            fixWebpackSourcePaths: true,
-            thresholds: {
-                statements: 80,
-                lines: 80,
-                branches: 80,
-                functions: 80
-            }
+            fixWebpackSourcePaths: true
+            // thresholds: {
+            //     statements: 80,
+            //     lines: 80,
+            //     branches: 80,
+            //     functions: 80
+            // }
         },
         reporters: ['progress', 'kjhtml'],
         junitReporter: {
