@@ -6,18 +6,16 @@
  * found under the LICENSE file in the root directory of this source tree.
  */
 
+// tslint:disable: no-any
+
 import { HttpResponse } from '@angular/common/http';
 
 import { CacheEntryOptions } from './cache-entry-options';
 
-export function handleCacheResponse<T>(response: HttpResponse<T>,
-    cacheInfoKey: string,
-    entryOptions: CacheEntryOptions): T;
+export function handleCacheResponse<T>(response: HttpResponse<T>, cacheInfoKey: string, entryOptions: CacheEntryOptions): T;
 
-export function handleCacheResponse(response: HttpResponse<any>,
-    cacheInfoKey: string,
-    entryOptions: CacheEntryOptions): any {
-    if (response.headers.has(cacheInfoKey)) {
+export function handleCacheResponse(response: HttpResponse<any>, cacheInfoKey: string, entryOptions: CacheEntryOptions): any {
+    if (cacheInfoKey && response.headers.has(cacheInfoKey)) {
         const headerValue = response.headers.get(cacheInfoKey);
         if (headerValue) {
             const cacheInfo = JSON.parse(headerValue) as CacheEntryOptions;
