@@ -6,18 +6,18 @@
  * found under the LICENSE file in the root directory of this source tree.
  */
 
-export * from './cache';
-export * from './cache.module';
-export * from './cache.service';
-export * from './cache-check-result';
 export * from './cache-entry-options';
 export * from './cache-item';
-export * from './cache-local-storage';
 export * from './cache-local-storage.module';
+export * from './cache-local-storage';
 export * from './cache-options';
+export * from './cache.module';
+export * from './cache.service';
+export * from './cache';
 export * from './handle-cache-response';
 export * from './initial-cache-data';
-export * from './memory-cache';
+export * from './logging-api';
 export * from './memory-cache.module';
+export * from './memory-cache';
 export * from './return-type';
 export * from './storage';
